# final_ims/backend/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stock Overflow API",
    "version": "1.0.0",
    "description": "Swagger documentation generated from the implemented NestJS controllers, DTOs, and TypeScript response types."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Application API base path"
    }
  ],
...
  "paths": {
...
    "/transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "operationId": "Transactions_findAll",
        "summary": "Find All",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          },
          {
            "name": "retailerId",
            "in": "query",
            "required": false,
            "description": "Query parameter: retailerId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "storeId",
            "in": "query",
            "required": false,
            "description": "Query parameter: storeId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customer",
            "in": "query",
            "required": false,
            "description": "Query parameter: customer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerEmail",
            "in": "query",
            "required": false,
            "description": "Query parameter: customerEmail",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "in": "query",
            "required": false,
            "description": "Query parameter: customerId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/common_database_types_TransactionRecord"
                  },
                  "type": "array"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "Transactions"
        ],
        "operationId": "Transactions_create",
        "summary": "Create",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common_database_types_TransactionRecord"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/transactions_dto_create_transaction_dto_CreateTransactionDto"
              }
            }
          }
        }
      }
    },
    "/transactions/latest": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "operationId": "Transactions_findLatest",
        "summary": "Find Latest",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          },
          {
            "name": "retailerId",
            "in": "query",
            "required": false,
            "description": "Query parameter: retailerId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "storeId",
            "in": "query",
            "required": false,
            "description": "Query parameter: storeId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customer",
            "in": "query",
            "required": false,
            "description": "Query parameter: customer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerEmail",
            "in": "query",
            "required": false,
            "description": "Query parameter: customerEmail",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "in": "query",
            "required": false,
            "description": "Query parameter: customerId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common_database_types_TransactionRecord"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/transactions/purchased-products": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "operationId": "Transactions_findPurchasedProducts",
        "summary": "Find Purchased Products",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          },
          {
            "name": "retailerId",
            "in": "query",
            "required": false,
            "description": "Query parameter: retailerId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "storeId",
            "in": "query",
            "required": false,
            "description": "Query parameter: storeId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customer",
            "in": "query",
            "required": false,
            "description": "Query parameter: customer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerEmail",
            "in": "query",
            "required": false,
            "description": "Query parameter: customerEmail",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "in": "query",
            "required": false,
            "description": "Query parameter: customerId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "emoji": {
                        "type": "string"
                      },
                      "lastOrderId": {
                        "type": "string"
                      },
                      "lastOrderedAt": {
                        "type": "string"
                      },
                      "name": {
                        "type": "string"
                      },
                      "productImg": {
                        "type": "string"
                      },
                      "sku": {
                        "type": "string"
                      },
                      "totalQty": {
                        "type": "number"
                      },
                      "totalSpent": {
                        "type": "number"
                      }
                    },
                    "required": [
                      "emoji",
                      "lastOrderId",
                      "lastOrderedAt",
                      "name",
                      "productImg",
                      "sku",
                      "totalQty",
                      "totalSpent"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/transactions/{orderId}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "operationId": "Transactions_findOne",
        "summary": "Find One",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          },
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "description": "Path parameter: orderId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "retailerId",
            "in": "query",
            "required": false,
            "description": "Query parameter: retailerId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "storeId",
            "in": "query",
            "required": false,
            "description": "Query parameter: storeId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customer",
            "in": "query",
            "required": false,
            "description": "Query parameter: customer",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerEmail",
            "in": "query",
            "required": false,
            "description": "Query parameter: customerEmail",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "in": "query",
            "required": false,
            "description": "Query parameter: customerId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common_database_types_TransactionRecord"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      },
      "patch": {
        "tags": [
          "Transactions"
        ],
        "operationId": "Transactions_update",
        "summary": "Update",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          },
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "description": "Path parameter: orderId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/common_database_types_TransactionRecord"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/transactions_dto_update_transaction_dto_UpdateTransactionDto"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Transactions"
        ],
        "operationId": "Transactions_remove",
        "summary": "Remove",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          },
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "description": "Path parameter: orderId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "item": {
                      "$ref": "#/components/schemas/common_database_types_TransactionRecord"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "item",
                    "message"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/transactions/reconcile": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "operationId": "Transactions_reconcileInventory",
        "summary": "Reconcile Inventory",
        "parameters": [
          {
            "$ref": "#/components/parameters/RoleHeader"
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "products": {
                      "type": "number"
                    },
                    "transactions": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "message",
                    "products",
                    "transactions"
                  ],
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "RoleHeader": {
        "name": "role",
        "in": "header",
        "required": false,
        "description": "Caller role context for this API. Supported values include `admin`, `retailer`, `supplier`, `consumer`, and `biller`.",
        "schema": {
          "type": "string",
          "enum": [
            "admin",
            "retailer",
            "supplier",
            "consumer",
            "biller"
          ]
        }
      }
    },
    "schemas": {
...
      "common_database_types_TransactionRecord": {
        "properties": {
          "coupon": {
            "type": "number"
          },
          "customer": {
            "type": "string"
          },
          "customerEmail": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "discount": {
            "type": "number"
          },
          "finalTotal": {
            "type": "number"
          },
          "items": {
            "items": {
              "properties": {
                "isReserved": {
                  "type": "boolean"
                },
                "name": {
                  "type": "string"
                },
                "price": {
                  "type": "number"
                },
                "quantity": {
                  "type": "number"
                },
                "requestIds": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "sku": {
                  "type": "string"
                },
                "total": {
                  "type": "number"
                }
              },
              "required": [
                "isReserved",
                "name",
                "price",
                "quantity",
                "requestIds",
                "sku",
                "total"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "orderId": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "retailerId": {
            "type": "string"
          },
          "roundoff": {
            "type": "number"
          },
          "shipping": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "store": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "subtotal": {
            "type": "number"
          },
          "tax": {
            "type": "number"
          },
          "timestamp": {
            "type": "string"
          }
        },
        "required": [
          "coupon",
          "customer",
          "discount",
          "finalTotal",
          "items",
          "orderId",
          "paymentMethod",
          "roundoff",
          "shipping",
          "status",
          "store",
          "storeId",
          "subtotal",
          "tax",
          "timestamp"
        ],
        "type": "object"
      },
...
      "transactions_dto_create_transaction_dto_CreateTransactionDto": {
        "properties": {
          "coupon": {
            "type": "number"
          },
          "customer": {
            "type": "string"
          },
          "customerEmail": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "discount": {
            "type": "number"
          },
          "items": {
            "items": {
              "properties": {
                "isReserved": {
                  "type": "boolean"
                },
                "name": {
                  "type": "string"
                },
                "price": {
                  "minimum": 0,
                  "type": "number"
                },
                "quantity": {
                  "minimum": 1,
                  "type": "number"
                },
                "requestIds": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "sku": {
                  "type": "string"
                },
                "total": {
                  "minimum": 0,
                  "type": "number"
                }
              },
              "required": [
                "name",
                "price",
                "quantity",
                "sku",
                "total"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "paymentMethod": {
            "type": "string"
          },
          "retailerId": {
            "type": "string"
          },
          "roundoff": {
            "type": "number"
          },
          "shipping": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "store": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "tax": {
            "type": "number"
          }
        },
        "required": [
          "customer",
          "items",
          "paymentMethod",
          "store"
        ],
        "type": "object"
      },
      "transactions_dto_update_transaction_dto_UpdateTransactionDto": {
        "properties": {
          "coupon": {
            "type": "number"
          },
          "customer": {
            "type": "string"
          },
          "customerEmail": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "discount": {
            "type": "number"
          },
          "items": {
            "items": {
              "properties": {
                "isReserved": {
                  "type": "boolean"
                },
                "name": {
                  "type": "string"
                },
                "price": {
                  "minimum": 0,
                  "type": "number"
                },
                "quantity": {
                  "minimum": 1,
                  "type": "number"
                },
                "requestIds": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "sku": {
                  "type": "string"
                },
                "total": {
                  "minimum": 0,
                  "type": "number"
                }
              },
              "required": [
                "name",
                "price",
                "quantity",
                "sku",
                "total"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "paymentMethod": {
            "type": "string"
          },
          "retailerId": {
            "type": "string"
          },
          "roundoff": {
            "type": "number"
          },
          "shipping": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "store": {
            "type": "string"
          },
          "storeId": {
            "type": "string"
          },
          "tax": {
            "type": "number"
          }
        },
        "type": "object"
      },
...
    }
  }
}